+++ src/backend/resolvers/mutations/organization/mbm/lbm/index.ts
@@ -1,9 +1,13 @@
 import createLbmOrganization from './create'
 import updateLbmOrganization from './update'
 import deleteLbmOrganization from './delete'
+import lbmTypeResolvers from './type'
+import lbmRelationalResolvers from './relational-resolvers'
 
 export default {
   createLbmOrganization,
   updateLbmOrganization,
   deleteLbmOrganization,
+  ...lbmTypeResolvers,
+  ...lbmRelationalResolvers,
 }