+++ src/backend/resolvers/mutations/organization/mbm/lbm/relational-resolvers/index.ts
@@ -1,0 +1,13 @@
+// import serviceAndServiceCategoryResolvers from './serviceAndServiceCategory'
+// import lbmAndServiceResolvers from './lbmAndService'
+// import lbmAndPersonResolvers from './lbmAndPerson'
+import lbmAndLbmTypeResolvers from './lbmAndLbmType'
+// import lbmAndKeyEventResolvers from './lbmAndKeyEvent'
+
+export default {
+  // ...serviceAndServiceCategoryResolvers,
+  // ...lbmAndServiceResolvers,
+  // ...lbmAndPersonResolvers,
+  ...lbmAndLbmTypeResolvers,
+  // ...lbmAndKeyEventResolvers,
+}