+++ src/backend/resolvers/mutations/organization/mbm/lbm/relational-resolvers/lbmAndLbmType/index.ts
@@ -1,0 +1,5 @@
+import connectLbmAndLbmType from './upsert'
+
+export default {
+  connectLbmAndLbmType,
+}