+++ src/backend/resolvers/mutations/organization/mbm/lbm/type/create.ts
@@ -1,0 +1,7 @@
+const createLbmType = (parent, { input }, { pulseCoreDb }, info) =>
+  pulseCoreDb
+    .collection('lbms.types')
+    .insertOne({ ...input })
+    .then(({ ops }) => ops[0])
+
+export default createLbmType