+++ src/backend/typeDefs/mutations/organization/mbm/lbm/type/create.ts
@@ -1,0 +1,10 @@
+import { gql } from 'apollo-server-express'
+
+const createLbmTypeTypeDefs = gql`
+  input CreateLbmTypeInput {
+    name: String!
+    description: String
+  }
+`
+
+export default createLbmTypeTypeDefs