+++ src/backend/typeDefs/mutations/organization/mbm/lbm/type/delete.ts
@@ -1,0 +1,9 @@
+import { gql } from 'apollo-server-express'
+
+const deleteLbmTypeTypeDefs = gql`
+  input DeleteLbmTypeInput {
+    _id: ID!
+  }
+`
+
+export default deleteLbmTypeTypeDefs